Implement StoryboardElementWithDuration
