+++ osu.Game/Storyboards/StoryboardAnimation.cs
@@ -7,7 +7,7 @@ using osu.Game.Storyboards.Drawables;
 
 namespace osu.Game.Storyboards
 {
-    public class StoryboardAnimation : StoryboardSprite
+    public class StoryboardAnimation : StoryboardElementWithDuration<DrawableStoryboardAnimation>
     {
         public int FrameCount;
         public double FrameDelay;
@@ -21,8 +21,7 @@ namespace osu.Game.Storyboards
             LoopType = loopType;
         }
 
-        public override Drawable CreateDrawable()
-            => new DrawableStoryboardAnimation(this);
+        public override DrawableStoryboardAnimation CreateStoryboardDrawable() => new DrawableStoryboardAnimation(this);
     }
 
     public enum AnimationLoopType
